failureinjection: catch runAsync panics as errors
Now that the failureinjection framework can spin up goroutines,
make sure to catch all panics as errors. This keeps it compatible
with the roachtest framework, where we don't want any uncaught
panics to take down an entire run.
