+++ pkg/roachprod/failureinjection/failures/BUILD.bazel
@@ -18,6 +18,7 @@ go_library(
         "//pkg/roachprod",
         "//pkg/roachprod/install",
         "//pkg/roachprod/logger",
+        "//pkg/roachprod/roachprodutil",
         "//pkg/util/retry",
         "//pkg/util/syncutil",
         "//pkg/util/sysutil",
