+++ pkg/roachprod/roachprodutil/BUILD.bazel
@@ -2,10 +2,15 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "roachprodutil",
-    srcs = ["identity.go"],
+    srcs = [
+        "identity.go",
+        "utils.go",
+    ],
     importpath = "github.com/cockroachdb/cockroach/pkg/roachprod/roachprodutil",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/roachprod/logger",
+        "//pkg/util/debugutil",
         "@com_github_cockroachdb_errors//:errors",
         "@com_google_cloud_go_storage//:storage",
         "@org_golang_google_api//idtoken",
